+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Function `wal-insert` now uses keyword arguments for
+  `allow-duplicates`, `before` and `quiet`.
+
 ## [v2.1.11]
 
 Longer safe.